Refactoring the filesystem project - Part 1 + Added new URI types
