+++ source/code/core/public/core/cexpr/stringid.hxx
@@ -3,12 +3,9 @@
 #include <fmt/format.h>
 
 
-namespace core
+namespace core::cexpr
 {
 
-
-namespace cexpr
-{
 
 //! \brief The hash type of the stringid type.
 enum class stringid_hash_type : uint64_t { };
@@ -18,13 +15,16 @@ enum class stringid_hash_type : uint64_t { };
 struct stringid_type
 {
     //! \brief The hash value.
-    stringid_hash_type hash_value{ 0 };
+    stringid_hash_type hash_value;
 
 #if CFG_RELEASE == 0
     //! \brief The origin string (or part of it).
-    char hash_origin[core::build::is_release ? 0 : 24]{ "" };
+    char hash_origin[24]{ "" };
 #endif
 };
+
+
+constexpr stringid_type stringid_invalid{ stringid_hash_type{ 0 } };
 
 
 #if CFG_RELEASE == 0
@@ -44,6 +44,9 @@ constexpr auto stringid_cexpr(std::string_view cstr) noexcept -> stringid_type
     return result;
 }
 
+//! \brief The argument type used to pass stringid_type values.
+using stringid_argument_type = const stringid_type&;
+
 #else
 
 //! \brief Create a new stringid_type value from the given string.
@@ -58,12 +61,26 @@ constexpr auto stringid_cexpr(std::string_view cstr) noexcept -> stringid_type
     return stringid(cstr);
 }
 
+//! \brief The argument type used to pass stringid_type values.
+using stringid_argument_type = stringid_type;
+
 #endif
 
-} // namespace cexpr
+
+//! \brief Equality operator for stringid_type values.
+bool operator==(stringid_argument_type left, stringid_argument_type right) noexcept
+{
+    return left.hash_value == right.hash_value;
+}
+
+//! \brief Inequality operator for stringid_type values.
+bool operator!=(stringid_argument_type left, stringid_argument_type right) noexcept
+{
+    return !(left == right);
+}
 
 
-} // namespace core
+} // namespace core::cexpr
 
 
 //////////////////////////////////////////////////////////////////////////
@@ -84,11 +101,18 @@ struct formatter<core::cexpr::stringid_type>
     template <typename FormatContext>
     auto format(const core::cexpr::stringid_type& strid, FormatContext& ctx)
     {
+        if (strid.hash_value == core::cexpr::stringid_hash_type{ 0 })
+        {
+            return fmt::format_to(ctx.begin(), "{{sid:<invalid>}}");
+        }
+        else
+        {
 #if CFG_RELEASE == 0
-        return fmt::format_to(ctx.begin(), "{{sid:{:16x}}} ['{}']", static_cast<std::underlying_type_t<decltype(strid.hash_value)>>(strid.hash_value), strid.hash_origin);
+            return fmt::format_to(ctx.begin(), "{{sid:{:16x}}}'{}'", static_cast<std::underlying_type_t<decltype(strid.hash_value)>>(strid.hash_value), strid.hash_origin);
 #else
-        return fmt::format_to(ctx.begin(), "{{sid:{:16x}}}", static_cast<std::underlying_type_t<decltype(strid.hash_value)>>(strid.hash_value));
+            return fmt::format_to(ctx.begin(), "{{sid:{:16x}}}", static_cast<std::underlying_type_t<decltype(strid.hash_value)>>(strid.hash_value));
 #endif
+        }
     }
 };
