+++ source/fbuild.bff
@@ -120,7 +120,26 @@ Settings
 
     .SolutionFolder =
     [
-        .Path      = 'core'
+        .Path      = '1. core'
+        .Projects = { }
+        ForEach( .Project in .VisualStudioProjects )
+        {
+            ^Projects + '$Project$-vcxproj'
+        }
+
+        ^SolutionProjects + .Projects
+    ]
+    ^SolutionFolders + .SolutionFolder
+    .VisualStudioProjects = { }
+
+
+    // Systems projects
+    //--------------------------------------------------------------------------
+    #include "code/filesystem/filesystem.bff"
+
+    .SolutionFolder =
+    [
+        .Path      = '2. systems'
         .Projects = { }
         ForEach( .Project in .VisualStudioProjects )
         {
